+++ pystalk/__init__.py
@@ -1,9 +1,14 @@
-from .client import BeanstalkClient, BeanstalkError
+from .client import BeanstalkClient, BeanstalkError, BeanstalkConnectionError
 from .pool import ProductionPool
 
-version_info = (0, 7, 0)
-__version__ = '.'.join(str(s) for s in version_info)
+__version__ = '0.8.0'
+
 __author__ = 'EasyPost <oss@easypost.com>'
 
 
-__all__ = ['BeanstalkClient', 'BeanstalkError', 'ProductionPool']
+__all__ = [
+    'BeanstalkClient',
+    'BeanstalkConnectionError',
+    'BeanstalkError',
+    'ProductionPool',
+]